Implemented RQ worker, moved tasks to their own module
