+++ webhaak.py
@@ -2,19 +2,18 @@ import binascii
 import json
 import logging
 import os
-import subprocess
-from datetime import datetime, timedelta
+from datetime import timedelta
 from functools import update_wrapper
-from multiprocessing import Process
 
-import git
-import pushover
 import strictyaml
 from flask import (Flask, Response, abort, current_app, jsonify, make_response,
                    request)
+from redis import Redis
+from rq import Queue
 from strictyaml import Bool, Map, MapPattern, Optional, Str
 
-import settings
+import tasks
+from tasks import settings
 
 app = Flask(__name__)
 app.debug = settings.DEBUG
@@ -52,243 +51,6 @@ schema = MapPattern(
 # Load the configuration of the various projects/hooks
 with open(settings.PROJECTS_FILE, 'r') as pf:
     projects = strictyaml.load(pf.read(), schema).data
-
-
-def notify_user(result, config):
-    """Send a PushOver message if configured, after git and command have run
-
-    result is a dictionary with fields:
-      command_result
-      status: 'OK' | 'error'
-      type: 'commanderror'
-      message
-    """
-    try:
-        triggerconfig = config[1]
-        projectname = '{}>{}'.format(config[0], triggerconfig['title'])
-        title = ''
-        branch = 'master'
-        command = 'n/a'
-        repo = 'n/a'
-        if 'command' in triggerconfig:
-            command = triggerconfig['command']
-        if 'branch' in triggerconfig:
-            branch = triggerconfig['branch']
-        if 'repo' in triggerconfig:
-            repo = triggerconfig['repo']
-        message = 'repo: {}\nbranch: {}\ncommand: {}\nruntime: {}'.format(
-            repo,
-            branch,
-            command,
-            result['runtime']
-        )
-        if result['status'] == 'OK':
-            title = "Hook for {} ran successfully".format(projectname)
-        else:
-            title = "Hook for {} failed: {}".format(projectname, result['type'])
-            message = message + '\n\n{}'.format(result['message'])
-        logging.debug(message)
-        logging.info('Sending notification...')
-        # TODO: option to send to Telegram chat
-        client = pushover.Client(settings.PUSHOVER_USERKEY, api_token=settings.PUSHOVER_APPTOKEN)
-        client.send_message(message, title=title)
-        logging.info('Notification sent')
-    except AttributeError:
-        logging.warning('Notification through PushOver failed because of missing configuration')
-
-
-def get_trigger_settings(appkey, triggerkey):
-    """Look up the trigger and return the repo and command to be updated and fired
-
-    :param appkey: application key part of the url
-    :param triggerkey: trigger key part of the url, sub part of the config
-    :return: tuple with project info and the trigger config
-    """
-    for project in projects:
-        if projects[project]['appkey'] == appkey:
-            for trigger in projects[project]['triggers']:
-                if projects[project]['triggers'][trigger]['triggerkey'] == triggerkey:
-                    triggerconfig = projects[project]['triggers'][trigger]
-                    triggerconfig['title'] = trigger
-                    return (project, triggerconfig)
-    return None
-
-
-def get_repo_basename(repo_url):
-    """Extract repository basename from its url, as that will be the name of directory it will be cloned into"""
-    result = os.path.basename(repo_url)
-    filename, file_extension = os.path.splitext(result)
-    if file_extension == '.git':
-        # Strip the .git from the name, as Git will do the same on non-bare checkouts
-        result = filename
-    return result
-
-
-def get_repo_version(repo_dir):
-    """Gets version of Git repo, based on latest tag, number of commits since, and latest commit hash
-
-    :param repo_dir: path to the Git repository
-    :return: string with version
-    """
-    # Make sure the working directory is our project
-    try:
-        version = subprocess.check_output(["git", "describe", "--always", "--tags"], stderr=None, cwd=repo_dir).strip()
-    except subprocess.CalledProcessError:
-        version = ''
-
-    try:
-        # byte string needs to be converted to a string
-        version = version.decode("utf-8")
-    except AttributeError:
-        # version already was a str
-        pass
-    return version
-
-
-def fetchinfo_to_str(fetchinfo):
-    """git.remote.FetchInfo to human readable representation"""
-    result = fetchinfo[0].note
-    return result
-
-
-def update_repo(config):
-    """Update (pull) the Git repo"""
-    projectname = config[0]
-    triggerconfig = config[1]
-
-    repo_url = triggerconfig['repo']
-    repo_parent = settings.REPOS_CACHE_DIR
-    if 'repoparent' in triggerconfig and triggerconfig['repoparent']:
-        repo_parent = triggerconfig['repoparent']
-
-    app.logger.info('[%s] Updating %s', projectname, repo_url)
-    app.logger.info('[%s] Repo parent %s', projectname, repo_parent)
-
-    # Ensure cache dir for webhaak exists and is writable
-    if not os.path.exists(repo_parent):
-        os.makedirs(repo_parent)  # throws OSError if repo_parent is not writable
-
-    # TODO: check whether dir exists with different repository
-    repo_dir = os.path.join(repo_parent, get_repo_basename(repo_url))
-    app.logger.info('[%s] Repo dir %s', projectname, repo_dir)
-    if os.path.isdir(repo_dir):
-        # Repo already exists locally, do a pull
-        app.logger.info('[%s] Repo exists, pull', projectname)
-
-        apprepo = git.Repo(repo_dir)
-        origin = apprepo.remote('origin')
-        result = fetchinfo_to_str(origin.fetch())  # assure we actually have data. fetch() returns useful information
-        app.logger.info('[%s] Fetch result: %s', projectname, result)
-    else:
-        # Repo needs to be cloned
-        app.logger.info('[%s] Repo does not exist yet, clone', projectname)
-        apprepo = git.Repo.init(repo_dir)
-        origin = apprepo.create_remote('origin', repo_url)
-        origin.fetch()                  # assure we actually have data. fetch() returns useful information
-        # Setup a local tracking branch of a remote branch
-        apprepo.create_head('master', origin.refs.master).set_tracking_branch(origin.refs.master)
-    branch = 'master'
-    if 'branch' in triggerconfig:
-        branch = triggerconfig['branch']
-    app.logger.info('[%s] checkout() branch \'%s\'', projectname, branch)
-    result = str(apprepo.git.checkout(branch))
-    # pull (so really update) the checked out branch to latest commit
-    origin.pull()
-    app.logger.info('[%s] Done pulling branch \'%s\'', projectname, branch)
-    return result
-
-
-def run_command(config, hook_info):
-    """Run the command(s) defined for this trigger"""
-    projectname = config[0]
-    triggerconfig = config[1]
-    if 'command' not in triggerconfig:
-        # No command to execute, return
-        app.logger.info('[%s] No command to execute', projectname)
-        return None
-    command = triggerconfig['command']
-    # Replace some placeholders to be used in executing scripts from one of the repos
-    repo_parent = settings.REPOS_CACHE_DIR
-    if 'repoparent' in triggerconfig and triggerconfig['repoparent']:
-        repo_parent = triggerconfig['repoparent']
-    if 'repo' in triggerconfig:
-        repo_url = triggerconfig['repo']
-        command = command.replace('REPODIR', os.path.join(repo_parent, get_repo_basename(repo_url)))
-    command = command.replace('CACHEDIR', settings.REPOS_CACHE_DIR)
-    if 'REPOVERSION' in command:
-        version = get_repo_version(os.path.join(repo_parent, projectname))
-        command = command.replace('REPOVERSION', version)
-
-    for key in hook_info:
-        if isinstance(hook_info[key], str):
-            command = command.replace(key.upper(), hook_info[key].replace('"', '\\"'))
-
-    command = command.strip()  # ensure no weird linefeeds and superfluous whitespace are there
-    app.logger.info('[%s] Executing `%s`', projectname, command)
-
-    # TODO: capture_output is new in Python 3.7, replaces stdout and stderr
-    # result = subprocess.run(command_parts, capture_output=True, check=True, shell=True, universal_newlines=True)
-    result = subprocess.run(
-        command,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        shell=True,
-        universal_newlines=True
-    )
-    return result
-
-
-def do_pull_andor_command(config, hook_info):
-    """Asynchronous task, performing the git pulling and the specified scripting inside a Process"""
-    projectname = config[0]
-    starttime = datetime.now()
-    result = {'application': projectname}
-    result['trigger'] = config[1]
-    if 'repo' in config[1]:
-        try:
-            result['repo_result'] = update_repo(config)
-            app.logger.info('[%s] result repo: %s', projectname, str(result['repo_result']))
-        except git.GitCommandError as e:
-            result = {'status': 'error', 'type': 'giterror', 'message': str(e)}
-            app.logger.error('[%s] giterror: %s', projectname, str(e))
-            result['runtime'] = datetime.now() - starttime
-            notify_user(result, config)
-            return
-        except (OSError, KeyError) as e:
-            result = {'status': 'error', 'type': 'oserror', 'message': str(e)}
-            app.logger.error('[%s] oserror: %s', projectname, str(e))
-            result['runtime'] = datetime.now() - starttime
-            notify_user(result, config)
-            return
-
-    cmdresult = run_command(config, hook_info)
-    if cmdresult and cmdresult.returncode == 0:
-        app.logger.info('[%s] success for command: %s', projectname, str(cmdresult.stdout))
-        result['status'] = 'OK'
-    elif not cmdresult:
-        app.logger.info('[%s] no command configured', projectname)
-        result['status'] = 'OK'
-    else:
-        result['status'] = 'error'
-        result['type'] = 'commanderror'
-        result['message'] = cmdresult.stderr.strip()
-        # TODO: seperate logfiles per job? Filename then based on appkey_triggerkey_timestamp.log
-        app.logger.error(
-            '[%s] commanderror with returncode %s: %s',
-            projectname,
-            str(cmdresult.returncode),
-            cmdresult.stderr
-        )
-        app.logger.error('[%s] stdout: %s', projectname, cmdresult.stdout)
-        app.logger.error('[%s] stderr: %s', projectname, cmdresult.stderr)
-
-    result['runtime'] = datetime.now() - starttime
-
-    if (
-            ('notify' not in config[1] or config[1]['notify'])
-            or (result['status'] == 'error' and ('notify_on_error' in config[1] and config[1]['notify_on_error']))
-    ):
-        notify_user(result, config)
 
 
 # == API request support functions/mixins ======
@@ -431,7 +193,7 @@ def apptrigger(appkey, triggerkey):
     :return: json Response
     """
     app.logger.info('%s on appkey: %s triggerkey: %s', request.method, appkey, triggerkey)
-    config = get_trigger_settings(appkey, triggerkey)
+    config = tasks.get_trigger_settings(appkey, triggerkey)
     if config is None:
         app.logger.error('appkey/triggerkey combo not found')
         return Response(json.dumps({'status': 'Error'}), status=404, mimetype='application/json')
@@ -634,14 +396,49 @@ def apptrigger(appkey, triggerkey):
         app.logger.debug(hook_info)
         app.logger.info(event_info)
 
-    p = Process(target=do_pull_andor_command, args=(config, hook_info,))
-    p.start()
+    #  print(config)
+    #  print('---')
+    #  print(hook_info)
+    #  print('------')
+
+    # Create RQ job (task) for this request
+    redis_conn = Redis()
+    q = Queue(connection=redis_conn)  # no args implies the default queue
+
+    # Delay execution of count_words_at_url('http://nvie.com')
+    # job = q.enqueue(tasks.do_pull_andor_command, args=(config, hook_info,))
+    job = q.enqueue("webhaak.tasks.do_pull_andor_command", args=(config, hook_info,))
+    app.logger.info('Enqueued job with id: %s' % job.id)
     return Response(
         json.dumps({
             'status': 'OK',
-            'message': 'Command accepted and will be run in the background'
+            'message': 'Command accepted and will be run in the background',
+            'job_id': job.id,
         }), status=200, mimetype='application/json'
     )
+
+
+@app.route('/status/<job_id>')
+def job_status(job_id):
+    """Show the status of job `job_id`
+
+    :param str job_id:
+    :return: dictionary with a task `status` and a `result`, including a relevant `message` on failure
+    :rtype: json
+    """
+    redis_conn = Redis()
+    q = Queue(connection=redis_conn)  # no args implies the default queue
+    job = q.fetch_job(job_id)
+    if job is None:
+        response = {'status': 'unknown'}
+    else:
+        response = {
+            'status': job.get_status(),
+            'result': job.result,
+        }
+        if job.is_failed:
+            response['message'] = job.exc_info.strip().split('\n')[-1]
+    return jsonify(response)
 
 
 @app.route('/monitor/monitor.html')
